Reduce the size of rate_limit fields

+++ src/builder/create_channel.rs
@@ -25,7 +25,7 @@ pub struct CreateChannel<'a> {
     #[serde(skip_serializing_if = "Option::is_none")]
     user_limit: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    rate_limit_per_user: Option<u64>,
+    rate_limit_per_user: Option<u16>,
     #[serde(skip_serializing_if = "Option::is_none")]
     position: Option<u32>,
     #[serde(skip_serializing_if = "Vec::is_empty")]
@@ -148,7 +148,7 @@ impl<'a> CreateChannel<'a> {
     /// [`MANAGE_MESSAGES`]: crate::model::permissions::Permissions::MANAGE_MESSAGES
     /// [`MANAGE_CHANNELS`]: crate::model::permissions::Permissions::MANAGE_CHANNELS
     #[doc(alias = "slowmode")]
-    pub fn rate_limit_per_user(mut self, seconds: u64) -> Self {
+    pub fn rate_limit_per_user(mut self, seconds: u16) -> Self {
         self.rate_limit_per_user = Some(seconds);
         self
     }
