+++ src/builder/edit_channel.rs
@@ -45,7 +45,7 @@ pub struct EditChannel<'a> {
     #[serde(skip_serializing_if = "Option::is_none")]
     nsfw: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    rate_limit_per_user: Option<u64>,
+    rate_limit_per_user: Option<u16>,
     #[serde(skip_serializing_if = "Option::is_none")]
     bitrate: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -67,7 +67,7 @@ pub struct EditChannel<'a> {
     #[serde(skip_serializing_if = "Option::is_none")]
     default_reaction_emoji: Option<Option<DefaultReaction>>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    default_thread_rate_limit_per_user: Option<u64>,
+    default_thread_rate_limit_per_user: Option<u16>,
     #[serde(skip_serializing_if = "Option::is_none")]
     default_sort_order: Option<SortOrder>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -181,7 +181,7 @@ impl<'a> EditChannel<'a> {
     /// [`MANAGE_MESSAGES`]: Permissions::MANAGE_MESSAGES
     /// [`MANAGE_CHANNELS`]: Permissions::MANAGE_CHANNELS
     #[doc(alias = "slowmode")]
-    pub fn rate_limit_per_user(mut self, seconds: u64) -> Self {
+    pub fn rate_limit_per_user(mut self, seconds: u16) -> Self {
         self.rate_limit_per_user = Some(seconds);
         self
     }
@@ -273,7 +273,7 @@ impl<'a> EditChannel<'a> {
     /// copied to the thread at creation time and does not live update.
     pub fn default_thread_rate_limit_per_user(
         mut self,
-        default_thread_rate_limit_per_user: u64,
+        default_thread_rate_limit_per_user: u16,
     ) -> Self {
         self.default_thread_rate_limit_per_user = Some(default_thread_rate_limit_per_user);
         self
